accounts/abi: move packNumbers into pack.go.
